Removed interface PermissionDAO
